feat: agent - eBPF Rename the file common.c to utils.c (#8876)
`utils.c` focuses on utility and practical functionality, containing small functional modules such as time, processes, and network interface configuration. Replacing `common.c` with `utils.c` is more appropriate in this context.
