+++ agent/src/ebpf/user/extended/extended.c
@@ -17,8 +17,8 @@
 #include <sys/stat.h>
 #include <bcc/perf_reader.h>
 #include "../config.h"
+#include "../common_utils.h"
 #include "../utils.h"
-#include "../common.h"
 #include "../mem.h"
 #include "../log.h"
 #include "../types.h"
