+++ agent/src/ebpf/user/load.c
@@ -26,7 +26,7 @@
 #include <stdlib.h>
 #include <limits.h>
 #include "config.h"
-#include "common.h"
+#include "utils.h"
 #include "log.h"
 #include "elf.h"
 #include <bcc/linux/bpf.h>
