+++ agent/src/ebpf/user/profile/perf_profiler.c
@@ -26,8 +26,8 @@
 #include <signal.h>		/* kill() */
 #include <bcc/perf_reader.h>
 #include "../config.h"
+#include "../common_utils.h"
 #include "../utils.h"
-#include "../common.h"
 #include "../mem.h"
 #include "../log.h"
 #include "../types.h"
